Mycelia Documentation

Welcome to the Mycelia documentation.

more

Installation

Install Julia (if not already installed)

I have had trouble getting the visualization libraries Plots.jl and Makie.jl (and associated packages) to load correctly on HPC due to the complexities of the default LDLIBRARYPATH

I imagine other research supercomputer users may have similar issues, although I don't have these issues on cloud vendors like GCP or AWS

To enable Julia to install all of it's own necessary dependencies independent of the system, I reset the LDLIBRARYPATH variable prior to launching Julia !!

This can be done easily when launching Julia from the command line by

export LD_LIBRARY_PATH="" && julia

And can be done for Julia jupyter kernels by setting the env key => value pair in the appropriate kernel.json file

Clone the repo directly

cd /path/where/you/want/the/repo
# for production usage
git clone https://github.com/cjprybol/Mycelia.git
# for development
git clone git@github.com:cjprybol/Mycelia.git

Or as Julia package

import Pkg
# for production usage
Pkg.add(url="https://github.com/cjprybol/Mycelia.git")
# for development
Pkg.develop(url="git@github.com:cjprybol/Mycelia.git")

Mycelia.add_bioconda_env — Method

add_bioconda_env(pkg; force) -> Union{Nothing, Base.Process}

Mycelia.add_edgemer_to_graph! — Method

add_edgemer_to_graph!(
    graph,
    record_identifier,
    index,
    observed_edgemer
) -> Any

A short description of the function

julia> 1 + 1
2

Mycelia.add_record_edgemers_to_graph! — Method

add_record_edgemers_to_graph!(graph) -> Any

A short description of the function

julia> 1 + 1
2

Mycelia.amino_acids_to_codons — Method

amino_acids_to_codons(

) -> Dict{BioSymbols.AminoAcid, DataType}

Mycelia.annotate_fasta — Method

annotate_fasta(
;
    fasta,
    identifier,
    basedir,
    mmseqsdb,
    threads
)

Mycelia.assess_alignment — Method

assess_alignment(
    a,
    b
) -> @NamedTuple{total_matches::Int64, total_edits::Int64}

Mycelia.assess_alignment_accuracy — Method

assess_alignment_accuracy(alignment_result) -> Any

Return proportion of matched bases in alignment to total matches + edits.

0-1, not %

Mycelia.assess_assembly_quality — Method

assess_assembly_quality(; assembly, observations, ks)

Mycelia.assess_dnamer_saturation — Method

assess_dnamer_saturation(
    fastx::AbstractString;
    power,
    outdir,
    min_k,
    max_k,
    threshold,
    kmers_to_assess
)

Mycelia.assess_dnamer_saturation — Method

assess_dnamer_saturation(
    fastxs::AbstractVector{<:AbstractString},
    kmer_type;
    kmers_to_assess,
    power,
    min_count
) -> Union{@NamedTuple{sampling_points::Vector{Int64}, unique_kmer_counts::Vector{Int64}}, NamedTuple{(:sampling_points, :unique_kmer_counts, :eof), <:Tuple{Vector, Vector{Int64}, Bool}}}

A short description of the function

julia> 1 + 1
2

Mycelia.assess_dnamer_saturation — Method

assess_dnamer_saturation(
    fastxs::AbstractVector{<:AbstractString};
    power,
    outdir,
    min_k,
    max_k,
    threshold,
    kmers_to_assess,
    plot
)

Mycelia.assess_optimal_kmer_alignment — Method

assess_optimal_kmer_alignment(
    kmer,
    observed_kmer
) -> Tuple{@NamedTuple{total_matches::Int64, total_edits::Int64}, Union{Missing, Bool}}

Used to determine which orientation provides an optimal alignment for initiating path likelihood analyses in viterbi analysis

Mycelia.bam_to_fastq — Method

bam_to_fastq(; bam, fastq)

Mycelia.bandage_visualize — Method

bandage_visualize(; gfa, img)

Mycelia.biosequences_to_counts_table — Method

biosequences_to_counts_table(; biosequences, k)

Mycelia.biosequences_to_dense_counts_table — Method

biosequences_to_dense_counts_table(; biosequences, k)

Mycelia.blastdb_to_fasta — Method

blastdb_to_fasta(; db, dbdir, compressed, outfile)

Mycelia.build_directed_kmer_graph — Method

build_directed_kmer_graph(; fastq, k, plot)

Mycelia.build_stranded_kmer_graph — Method

build_stranded_kmer_graph(
    kmer_type,
    observations::AbstractVector{<:Union{FASTX.FASTA.Record, FASTX.FASTQ.Record}}
) -> MetaGraphs.MetaDiGraph{T, Float64} where T<:Integer

Create a weighted, strand-specific kmer (de bruijn) graph from a set of kmers and a series of sequence observations in FASTA format.

Mycelia.canonicalize_kmer_counts! — Method

canonicalize_kmer_counts!(kmer_counts) -> Any

Mycelia.canonicalize_kmer_counts — Method

canonicalize_kmer_counts(kmer_counts) -> Any

Mycelia.codon_optimize — Method

codon_optimize(
;
    normalized_codon_frequencies,
    protein_sequence,
    n_iterations
)

Mycelia.codons_to_amino_acids — Method

codons_to_amino_acids(

) -> Union{Dict{_A, BioSequences.LongAA} where _A, Dict{_A, V} where {_A, V<:(Kmers.Kmer{BioSequences.AminoAcidAlphabet, _A, _B} where {_B, _A})}}

Mycelia.concatenate_files — Method

concatenate_files(; files, file)

Join fasta files without any regard to record uniqueness.

A cross-platform version of cat *.fasta > joint.fasta

See mergefastafiles

Mycelia.contig_is_circular — Method

contig_is_circular(
    graph_file::String,
    contig_name::String
) -> Any

Returns bool indicating whether the contig is a circle

graphfile = path to assembly graph.gfa file contigname = name of the contig

julia> 1 + 1
2

Mycelia.contig_is_cleanly_assembled — Method

contig_is_cleanly_assembled(
    graph_file::String,
    contig_name::String
) -> Bool

Returns bool indicating whether the contig is cleanly assembled.

By cleanly assembled we mean that the contig does not have other contigs attached in the same connected component.

graphfile = path to assembly graph.gfa file contigname = name of the contig

julia> 1 + 1
2

Mycelia.convert — Method

convert(args)

Mycelia.copy_with_unique_identifier — Method

copy_with_unique_identifier(
    infile,
    out_directory,
    unique_identifier;
    force
) -> Any

Mycelia.count_canonical_kmers — Method

count_canonical_kmers(_::Type{KMER_TYPE}, sequences) -> Any

Mycelia.count_kmers — Method

count_kmers(
    _::Type{KMER_TYPE},
    fastx_file::AbstractString
) -> Any

Mycelia.count_kmers — Method

count_kmers(
    _::Type{Kmers.Kmer{A<:BioSequences.AminoAcidAlphabet, K}},
    sequence::BioSequences.LongSequence
) -> OrderedCollections.OrderedDict{K, Int64} where K<:(Kmers.Kmer{BioSequences.AminoAcidAlphabet, _A, _B} where {_B, _A})

Mycelia.count_kmers — Method

count_kmers(
    _::Type{Kmers.Kmer{A<:BioSequences.DNAAlphabet, K}},
    sequence::BioSequences.LongSequence
) -> Any

Mycelia.count_kmers — Method

count_kmers(
    _::Type{KMER_TYPE},
    sequences::Union{FASTX.FASTA.Reader, FASTX.FASTQ.Reader}
) -> Any

Mycelia.count_kmers — Method

count_kmers(
    _::Type{KMER_TYPE},
    record::Union{FASTX.FASTA.Record, FASTX.FASTQ.Record}
) -> Any

Mycelia.count_kmers — Method

count_kmers(
    _::Type{KMER_TYPE},
    fastx_files::AbstractArray{T<:AbstractString, 1}
) -> Any

Mycelia.count_kmers — Method

count_kmers(
    _::Type{KMER_TYPE},
    records::AbstractArray{T<:Union{FASTX.FASTA.Record, FASTX.FASTQ.Record}, 1}
) -> Any

Mycelia.count_matrix_to_probability_matrix — Method

count_matrix_to_probability_matrix(counts_matrix) -> Any

Mycelia.count_records — Method

count_records(fastx) -> Int64

Mycelia.countmap_columns — Method

countmap_columns(table)

Description

julia> 1 + 1
2

Mycelia.create_database — Method

create_database(; database, address, username, password)

A short description of the function

julia> 1 + 1
2

Mycelia.create_node_constraints — Method

create_node_constraints(
    graph;
    address,
    username,
    password,
    database
)

Description

julia> 1 + 1
2

Mycelia.create_tarchive — Method

create_tarchive(; directory, tarchive)

Mycelia.cypher — Method

cypher(
    cmd;
    address,
    username,
    password,
    format,
    database
) -> Cmd

Description

julia> 1 + 1
2

Mycelia.deduplicate_fasta_file — Method

deduplicate_fasta_file(in_fasta, out_fasta) -> Any

Mycelia.determine_fasta_coverage — Method

determine_fasta_coverage(bam) -> Any

Mycelia.determine_max_canonical_kmers — Method

determine_max_canonical_kmers(k, ALPHABET) -> Any

Mycelia.determine_max_possible_kmers — Method

determine_max_possible_kmers(k, ALPHABET) -> Any

Mycelia.determine_primary_contig — Method

determine_primary_contig(qualimap_results) -> Any

Determines the contig with the greatest number of total bases mapping to it

julia> 1 + 1
2

Mycelia.determine_read_lengths — Method

determine_read_lengths(
    fastq_file;
    total_reads
) -> Vector{Int64}

Mycelia.distance_matrix_to_newick — Method

distance_matrix_to_newick(
;
    distance_matrix,
    labels,
    outfile
)

Create distance matrix from a column-major counts matrix (features as rows and entities as columns) where distance is a proportional to total feature count magnitude (size) and cosine similarity (relative frequency)

Mycelia.document_frequency — Method

document_frequency(documents) -> Dict{_A, Int64} where _A

A short description of the function

julia> 1 + 1
2

Mycelia.download_and_filter_sra_reads — Method

download_and_filter_sra_reads(
;
    outdir,
    srr_identifier
) -> Union{Nothing, Bool}

Mycelia.download_bandage — Function

download_bandage() -> String
download_bandage(outdir) -> Any

Mycelia.download_blast_db — Method

download_blast_db(; db, dbdir, source, wait)

Smart downloading of blast dbs depending on interactive, non interactive context

For a list of all available databases, run: Mycelia.list_blastdbs()

Mycelia.download_genome_by_accession — Method

download_genome_by_accession(
;
    accession,
    outdir,
    compressed
)

Mycelia.download_genome_by_ftp — Method

download_genome_by_ftp(; ftp, outdir)

Mycelia.download_mmseqs_db — Method

download_mmseqs_db(; db, dbdir, force, wait)

Download mmseqs databases

  Name                  Type            Taxonomy        Url                                                           
- UniRef100             Aminoacid            yes        https://www.uniprot.org/help/uniref
- UniRef90              Aminoacid            yes        https://www.uniprot.org/help/uniref
- UniRef50              Aminoacid            yes        https://www.uniprot.org/help/uniref
- UniProtKB             Aminoacid            yes        https://www.uniprot.org/help/uniprotkb
- UniProtKB/TrEMBL      Aminoacid            yes        https://www.uniprot.org/help/uniprotkb
- UniProtKB/Swiss-Prot  Aminoacid            yes        https://uniprot.org
- NR                    Aminoacid            yes        https://ftp.ncbi.nlm.nih.gov/blast/db/FASTA
- NT                    Nucleotide             -        https://ftp.ncbi.nlm.nih.gov/blast/db/FASTA
- GTDB                  Aminoacid            yes        https://gtdb.ecogenomic.org
- PDB                   Aminoacid              -        https://www.rcsb.org
- PDB70                 Profile                -        https://github.com/soedinglab/hh-suite
- Pfam-A.full           Profile                -        https://pfam.xfam.org
- Pfam-A.seed           Profile                -        https://pfam.xfam.org
- Pfam-B                Profile                -        https://xfam.wordpress.com/2020/06/30/a-new-pfam-b-is-released
- CDD                   Profile                -        https://www.ncbi.nlm.nih.gov/Structure/cdd/cdd.shtml
- eggNOG                Profile                -        http://eggnog5.embl.de
- VOGDB                 Profile                -        https://vogdb.org
- dbCAN2                Profile                -        http://bcb.unl.edu/dbCAN2
- SILVA                 Nucleotide           yes        https://www.arb-silva.de
- Resfinder             Nucleotide             -        https://cge.cbs.dtu.dk/services/ResFinder
- Kalamari              Nucleotide           yes        https://github.com/lskatz/Kalamari

Mycelia.drop_empty_columns! — Method

drop_empty_columns!(
    df::DataFrames.AbstractDataFrame
) -> DataFrames.AbstractDataFrame

Identify all columns that have only missing or empty values, and remove those columns from the dataframe in-place.

Returns a modified version of the original dataframe.

See also: dropemptycolumns

Mycelia.drop_empty_columns — Method

drop_empty_columns(df::DataFrames.AbstractDataFrame) -> Any

Identify all columns that have only missing or empty values, and remove those columns from the dataframe.

Returns a modified copy of the dataframe.

See also: dropemptycolumns!

Mycelia.edge_path_to_sequence — Method

edge_path_to_sequence(kmer_graph, edge_path) -> Any

Mycelia.edge_probability — Method

edge_probability(stranded_kmer_graph, edge) -> Any

Mycelia.edgemer_to_vertex_kmers — Method

edgemer_to_vertex_kmers(
    edgemer
) -> Tuple{Kmers.Kmer{BioSequences.DNAAlphabet{2}}, Kmers.Kmer{BioSequences.DNAAlphabet{2}}}

Mycelia.equivalent_fasta_sequences — Method

equivalent_fasta_sequences(fasta_1, fasta_2) -> Bool

Mycelia.error_rate_to_q_value — Method

error_rate_to_q_value(error_rate) -> Any

Mycelia.export_blast_db — Method

export_blast_db(; path_to_db, fasta)

Mycelia.export_blast_db_taxonomy_table — Method

export_blast_db_taxonomy_table(; path_to_db, outfile)

Mycelia.fasta_and_gff_to_genbank — Method

fasta_and_gff_to_genbank(; fasta, gff, genbank)

Mycelia.fasta_genome_size — Method

fasta_genome_size(fasta_file) -> Any

Mycelia.fasta_list_to_dense_counts_table — Method

fasta_list_to_dense_counts_table(; fasta_list, k, alphabet)

Create a dense kmer counts table (canonical for DNA, stranded for RNA & AA) for each fasta provided in a list. Scales very well for large numbers of organisms/fasta files, but not for k. Recommended for k <= 13, although 17 may still be possible

Mycelia.fasta_table_to_fasta — Method

fasta_table_to_fasta(fasta_df) -> Any

Mycelia.fasta_to_reference_kmer_counts — Method

fasta_to_reference_kmer_counts(; kmer_type, fasta)

Mycelia.fasta_to_table — Method

fasta_to_table(fasta) -> DataFrames.DataFrame

Mycelia.fasta_xam_mapping_stats — Method

fasta_xam_mapping_stats(; fasta, xam)

Mycelia.fastani_list — Method

fastani_list(
;
    query_list,
    reference_list,
    outfile,
    threads,
    force
) -> Union{Nothing, Base.Process}

Run fastani with a query and reference list

julia> 1 + 1
2

Mycelia.fasterq_dump — Method

fasterq_dump(
;
    outdir,
    srr_identifier
) -> Union{Nothing, Base.Process}

Mycelia.fastq_record — Method

fastq_record(; identifier, sequence, quality_scores)

Mycelia.fastx_stats — Method

fastx_stats(fastq) -> Base.Process

Mycelia.fastx_to_contig_lengths — Method

fastx_to_contig_lengths(
    fastx
) -> OrderedCollections.OrderedDict

Mycelia.fastx_to_kmer_graph — Method

fastx_to_kmer_graph(
    KMER_TYPE,
    fastxs::AbstractVector{<:AbstractString}
) -> MetaGraphs.MetaGraph

Create an in-memory kmer-graph that records:

all kmers

counts

all observed edges between kmers

edge orientations

edge counts

julia> 1 + 1
2

Mycelia.fibonacci_numbers_less_than — Method

fibonacci_numbers_less_than(
    n::Int64
) -> Union{Vector{Any}, Vector{Int64}}

Mycelia.filesize_human_readable — Method

filesize_human_readable(f) -> Any

Gets the size of a file and returns it in a human-readable format.

Arguments

f: The path to the file, either as a String or an AbstractString.

Returns

A string representing the file size in a human-readable format (e.g., "3.40 MB").

Details

This function internally uses filesize(f) to get the file size in bytes, then leverages Base.format_bytes to convert it into a human-readable format with appropriate units (KB, MB, GB, etc.).

Examples

julia> filesize_human_readable("my_image.jpg")
"2.15 MB"

See Also

filesize: Gets the size of a file in bytes.

Base.format_bytes: Converts a byte count into a human-readable string.

Mycelia.filter_long_reads — Method

filter_long_reads(
;
    in_fastq,
    out_fastq,
    min_mean_q,
    keep_percent
)

Filter and process long reads from a FASTQ file using Filtlong.

This function filters long sequencing reads based on quality and length criteria, then compresses the output using pigz.

Arguments

in_fastq::String: Path to the input FASTQ file.

out_fastq::String: Path to the output filtered and compressed FASTQ file. Defaults to the input filename with ".filtlong.fq.gz" appended.

min_mean_q::Int: Minimum mean quality score for reads to be kept. Default is 20.

keep_percent::Int: Percentage of reads to keep after filtering. Default is 95.

Returns

Cmd: A pipeline command that can be run to execute the filtering and compression.

Details

This function uses Filtlong to filter long reads and pigz for compression. It requires the Bioconda environment for Filtlong to be set up, which is handled internally.

Example

filter_cmd = filter_long_reads(
    in_fastq = "input.fastq.gz",
    out_fastq = "filtered_output.fq.gz",
    min_mean_q = 25,
    keep_percent = 90
)
run(filter_cmd)

Mycelia.find_matching_prefix — Method

find_matching_prefix(
    filename1::String,
    filename2::String;
    strip_trailing_delimiters
) -> String

Mycelia.find_nonempty_columns — Method

find_nonempty_columns(df) -> Any

Identify all columns that have only missing or empty values

Returns as a bit array

See also: dropemptycolumns, dropemptycolumns!

Mycelia.find_resampling_stretches — Method

find_resampling_stretches(
;
    record_kmer_solidity,
    solid_branching_kmer_indices
)

Mycelia.fit_optimal_number_of_clusters — Function

fit_optimal_number_of_clusters(
    distance_matrix
) -> @NamedTuple{optimal_number_of_clusters::Int64, ks_assessed::Vector{Int64}, within_cluster_sum_of_squares::Vector{Float64}, silhouette_scores::Vector{Float64}}
fit_optimal_number_of_clusters(
    distance_matrix,
    ks_to_try
) -> NamedTuple{(:optimal_number_of_clusters, :ks_assessed, :within_cluster_sum_of_squares, :silhouette_scores), <:Tuple{Any, Any, Vector{Float64}, Vector{Float64}}}

A short description of the function

julia> 1 + 1
2

Mycelia.fit_optimal_number_of_clusters_hclust — Function

fit_optimal_number_of_clusters_hclust(
    distance_matrix
) -> NamedTuple{(:optimal_number_of_clusters, :ks_assessed, :silhouette_scores, :hclust_result), <:Tuple{Int64, Vector{Int64}, Vector{Float64}, Clustering.Hclust}}
fit_optimal_number_of_clusters_hclust(
    distance_matrix,
    ks_to_try
) -> NamedTuple{(:optimal_number_of_clusters, :ks_assessed, :silhouette_scores, :hclust_result), <:Tuple{Any, Any, Vector{Float64}, Clustering.Hclust}}

A short description of the function

julia> 1 + 1
2

Mycelia.frequency_matrix_to_cosine_distance_matrix — Method

frequency_matrix_to_cosine_distance_matrix(
    probability_matrix
) -> Any

Create cosine distance matrix from a column-major counts matrix (features as rows and entities as columns) where distance is a proportional to cosine similarity (relative frequency)

Mycelia.frequency_matrix_to_euclidean_distance_matrix — Method

frequency_matrix_to_euclidean_distance_matrix(
    counts_table
) -> Any

Create a Euclidean distance matrix from a column-major counts matrix (features as rows and entities as columns), where distance is proportional to total feature count magnitude (size).

Mycelia.genbank_to_codon_frequencies — Method

genbank_to_codon_frequencies(
    genbank;
    allow_all
) -> Dict{BioSymbols.AminoAcid, Dict{Kmers.Kmer{BioSequences.DNAAlphabet{2}, 3, 1}, Int64}}

Mycelia.genbank_to_fasta — Method

genbank_to_fasta(; genbank, fasta, force)

Mycelia.generate_all_possible_canonical_kmers — Method

generate_all_possible_canonical_kmers(k, alphabet) -> Any

Create distance matrix from a column-major counts matrix (features as rows and entities as columns) where distance is a proportional to total feature count magnitude (size) and cosine similarity (relative frequency)

Mycelia.generate_all_possible_kmers — Method

generate_all_possible_kmers(k, alphabet) -> Any

Create distance matrix from a column-major counts matrix (features as rows and entities as columns) where distance is a proportional to total feature count magnitude (size) and cosine similarity (relative frequency)

Mycelia.generate_transterm_coordinates_from_fasta — Method

generate_transterm_coordinates_from_fasta(fasta) -> Any

Mycelia.generate_transterm_coordinates_from_gff — Method

generate_transterm_coordinates_from_gff(gff_file) -> Any

Mycelia.get_genbank — Method

get_genbank(
;
    db,
    accession,
    ftp
) -> Union{Nothing, GenomicAnnotations.GenBank.Reader}

Get dna (db = "nuccore") or protein (db = "protein") sequences from NCBI or get fasta directly from FTP site

Mycelia.get_gff — Method

get_gff(; db, accession, ftp) -> Any

Get dna (db = "nuccore") or protein (db = "protein") sequences from NCBI or get fasta directly from FTP site

Mycelia.get_kmer_index — Method

get_kmer_index(kmers, kmer) -> Any

Mycelia.get_sequence — Method

get_sequence(
;
    db,
    accession,
    ftp
) -> Union{Nothing, FASTX.FASTA.Reader}

Get dna (db = "nuccore") or protein (db = "protein") sequences from NCBI or get fasta directly from FTP site

julia> 1 + 1
2

Mycelia.gfa_to_fasta — Method

gfa_to_fasta(; gfa, fasta)

Mycelia.gfa_to_structure_table — Method

gfa_to_structure_table(
    gfa
) -> NamedTuple{(:contig_table, :records), <:Tuple{DataFrames.DataFrame, Any}}

Mycelia.githash — Method

githash(; short) -> SubString{String}

Mycelia.graph_to_gfa — Method

graph_to_gfa(; graph, outfile)

Mycelia.install_hashdeep — Method

install_hashdeep() -> Union{Nothing, Base.Process}

Mycelia.is_equivalent — Method

is_equivalent(a, b) -> Any

Mycelia.isolate_normalized_primary_contig — Method

isolate_normalized_primary_contig(
    assembled_fasta,
    assembled_gfa,
    qualimap_report_txt,
    identifier,
    k::Int64;
    primary_contig_fasta
) -> String

Primary contig is defined as the contig with the most bases mapped to it

In the context of picking out phage from metagenomic assemblies the longest contig is often bacteria whereas the highest coverage contigs are often primer-dimers or other PCR amplification artifacts.

Taking the contig that has the most bases mapped to it as a product of length * depth is cherry picked as our phage

Mycelia.iterative_polishing — Function

iterative_polishing(
    fastq
) -> Vector{T} where T<:(NamedTuple{(:fastq, :k), <:Tuple{Any, Any}})
iterative_polishing(
    fastq,
    max_k
) -> Vector{T} where T<:(NamedTuple{(:fastq, :k), <:Tuple{Any, Any}})
iterative_polishing(
    fastq,
    max_k,
    plot
) -> Vector{T} where T<:(NamedTuple{(:fastq, :k), <:Tuple{Any, Any}})

Mycelia.jaccard_distance — Method

jaccard_distance(set1, set2) -> Any

Mycelia.jaccard_similarity — Method

jaccard_similarity(set1, set2) -> Any

Mycelia.jellyfish_count — Method

jellyfish_count(
;
    fastx,
    k,
    threads,
    max_mem,
    canonical,
    outfile,
    conda_check
)

Mycelia.jellyfish_counts_to_kmer_frequency_histogram — Function

jellyfish_counts_to_kmer_frequency_histogram(
    jellyfish_counts_file
) -> Any
jellyfish_counts_to_kmer_frequency_histogram(
    jellyfish_counts_file,
    outfile
) -> Any

Mycelia.jitter — Method

jitter(x, n) -> Any

Mycelia.kmer_counts_dict_to_vector — Method

kmer_counts_dict_to_vector(
    kmer_to_index_map,
    kmer_counts
) -> Any

Mycelia.kmer_counts_to_cosine_similarity — Method

kmer_counts_to_cosine_similarity(
    kmer_counts_1,
    kmer_counts_2
) -> Any

Mycelia.kmer_counts_to_js_divergence — Method

kmer_counts_to_js_divergence(
    kmer_counts_1,
    kmer_counts_2
) -> Any

Mycelia.kmer_counts_to_merqury_qv — Method

kmer_counts_to_merqury_qv(
;
    raw_data_counts,
    assembly_counts
)

Mycelia.kmer_path_to_sequence — Method

kmer_path_to_sequence(kmer_path) -> Any

Mycelia.ks — Method

ks(; min, max) -> Vector{Int64}

Mycelia.lawrencium_sbatch — Method

lawrencium_sbatch(
;
    job_name,
    mail_user,
    mail_type,
    logdir,
    partition,
    qos,
    account,
    nodes,
    ntasks,
    time,
    cpus_per_task,
    mem_gb,
    cmd
)

Submit a command to SLURM using sbatch

Mycelia.list_blastdbs — Method

list_blastdbs(; source) -> Vector{String}

Mycelia.list_classes — Method

list_classes() -> DataFrames.DataFrame

Mycelia.list_databases — Method

list_databases(; address, username, password)

A short description of the function

julia> 1 + 1
2

Mycelia.list_families — Method

list_families() -> DataFrames.DataFrame

Mycelia.list_full_taxonomy — Method

list_full_taxonomy() -> DataFrames.DataFrame

Mycelia.list_genera — Method

list_genera() -> DataFrames.DataFrame

Mycelia.list_kingdoms — Method

list_kingdoms() -> DataFrames.DataFrame

Mycelia.list_orders — Method

list_orders() -> DataFrames.DataFrame

Mycelia.list_phylums — Method

list_phylums() -> DataFrames.DataFrame

Mycelia.list_rank — Method

top

superkingdom/domain

kingdom

phylum

class

order

family

genus

species

Mycelia.list_ranks — Method

list_ranks(; synonyms) -> Vector{String}

Mycelia.list_species — Method

list_species() -> DataFrames.DataFrame

Mycelia.list_subtaxa — Method

list_subtaxa(taxid) -> Vector{Int64}

Mycelia.list_superkingdoms — Method

list_superkingdoms() -> DataFrames.DataFrame

Mycelia.list_toplevel — Method

list_toplevel() -> DataFrames.DataFrame

Mycelia.load_blast_db_taxonomy_table — Method

load_blast_db_taxonomy_table(
    compressed_blast_db_taxonomy_table_file
) -> DataFrames.DataFrame

Mycelia.load_genbank_metadata — Method

load_genbank_metadata() -> DataFrames.DataFrame

Description

julia> 1 + 1
2

Mycelia.load_graph — Method

load_graph(file) -> Any

Mycelia.load_graph — Method

load_graph(file::String) -> Any

Mycelia.load_jellyfish_counts — Method

load_jellyfish_counts(
    jellyfish_counts
) -> DataFrames.DataFrame

Mycelia.load_ncbi_metadata — Method

load_ncbi_metadata(db) -> DataFrames.DataFrame

Description

julia> 1 + 1
2

Mycelia.load_refseq_metadata — Method

load_refseq_metadata() -> DataFrames.DataFrame

Description

julia> 1 + 1
2

Mycelia.local_blast_database_info — Method

local_blast_database_info(; blastdbs_dir) -> Any

Mycelia.merge_fasta_files — Method

merge_fasta_files(; fasta_files, fasta_file)

Join fasta files while adding origin prefixes to the identifiers.

Does not guarantee uniqueness but will warn if conflicts arise

Mycelia.minimap_index — Method

minimap_index(
;
    fasta,
    mem_gb,
    mapping_type,
    threads,
    as_string,
    denominator
)

Run this on the machine you intend to use to map the reads to confirm the index will fit

Mycelia.minimap_map — Method

minimap_map(
;
    fasta,
    fastq,
    mapping_type,
    as_string,
    mem_gb,
    threads,
    denominator
)

aligning and compressing. No sorting or filtering.

Use shell_only=true to get string command to submit to SLURM

Mycelia.minimap_map_paired_end_with_index — Method

minimap_map_paired_end_with_index(
;
    fasta,
    forward,
    reverse,
    mem_gb,
    threads,
    outdir,
    as_string,
    mapping_type,
    denominator
)

Mycelia.minimap_map_with_index — Method

minimap_map_with_index(
;
    fasta,
    mem_gb,
    mapping_type,
    threads,
    fastq,
    as_string,
    denominator
)

Mycelia.mmseqs_pairwise_search — Method

mmseqs_pairwise_search(; fasta, output)

Mycelia.mutate_sequence — Method

mutate_sequence(reference_sequence) -> Tuple{Any, Any}

Mycelia.name2taxid — Method

name2taxid(name) -> DataFrames.DataFrame

Mycelia.names2taxids — Method

names2taxids(names::AbstractVector{<:AbstractString}) -> Any

Mycelia.ncbi_ftp_path_to_url — Method

ncbi_ftp_path_to_url(; ftp_path, extension)

Extensions include:

genomic.fna.gz

genomic.gff.gz

protein.faa.gz

assembly_report.txt

assembly_stats.txt

cdsfromgenomic.fna.gz

feature_count.txt.gz

feature_table.txt.gz

genomic.gbff.gz

genomic.gtf.gz

protein.gpff.gz

translated_cds.faa.gz

julia> 1 + 1
2

Mycelia.ncbi_genome_download_accession — Method

function ncbi_genome_download_accession(;
        accession,
        outdir = pwd(),
        outpath = joinpath(outdir, accession * ".zip"),
        include_string = "genome"
    )

Download an accession using NCBI datasets command line tool

the .zip download output to outpath will be unzipped

returns the outfolder

ncbi's default include string is include_string = "gff3,rna,cds,protein,genome,seq-report"

Mycelia.ncbi_taxon_summary — Method

ncbi_taxon_summary(taxa_id) -> DataFrames.DataFrame

Mycelia.nearest_prime — Method

nearest_prime(n::Int64) -> Int64

Mycelia.nersc_sbatch — Method

nersc_sbatch(
;
    job_name,
    mail_user,
    mail_type,
    logdir,
    scriptdir,
    qos,
    nodes,
    ntasks,
    time,
    cpus_per_task,
    mem_gb,
    cmd,
    constraint
)

Submit a command to SLURM using sbatch

https://docs.nersc.gov/jobs/policy/ https://docs.nersc.gov/systems/perlmutter/architecture/#cpu-nodes

default is to use shared qos

use

regular

preempt (reduced credit usage but not guaranteed to finish)

premium (priorty runs limited to 5x throughput)

https://docs.nersc.gov/systems/perlmutter/running-jobs/#tips-and-tricks

Mycelia.nersc_sbatch_shared — Method

nersc_sbatch_shared(
;
    job_name,
    mail_user,
    mail_type,
    logdir,
    qos,
    nodes,
    ntasks,
    time,
    cpus_per_task,
    mem_gb,
    cmd,
    constraint
)

Submit a command to SLURM using sbatch

https://docs.nersc.gov/jobs/policy/ https://docs.nersc.gov/systems/perlmutter/architecture/#cpu-nodes

default is to use shared qos

use

regular

preempt (reduced credit usage but not guaranteed to finish)

premium (priorty runs limited to 5x throughput)

max request is 512Gb memory and 128 cores per node

https://docs.nersc.gov/systems/perlmutter/running-jobs/#tips-and-tricks

Mycelia.node_type_to_dataframe — Method

node_type_to_dataframe(; node_type, graph)

Mycelia.normalize_codon_frequencies — Method

normalize_codon_frequencies(
    codon_frequencies
) -> Dict{BioSymbols.AminoAcid, Dict{Kmers.Kmer{BioSequences.DNAAlphabet{2}, 3, 1}, Float64}}

Mycelia.normalize_countmap — Method

normalize_countmap(countmap) -> Dict

Mycelia.normalize_distance_matrix — Method

normalize_distance_matrix(distance_matrix) -> Any

Create distance matrix from a column-major counts matrix (features as rows and entities as columns) where distance is a proportional to total feature count magnitude (size) and cosine similarity (relative frequency)

Mycelia.normalize_vcf — Method

normalize_vcf(; reference_fasta, vcf_file)

Mycelia.normalized_current_datetime — Method

normalized_current_datetime() -> String

Mycelia.observe — Method

observe(
    record::Union{FASTX.FASTA.Record, FASTX.FASTQ.Record};
    error_rate
)

Mycelia.observe — Method

observe(
    records::AbstractArray{R<:Union{FASTX.FASTA.Record, FASTX.FASTQ.Record}, 1};
    weights,
    N,
    outfile,
    error_rate
) -> String

Mycelia.observe — Method

observe(
    sequence::BioSequences.LongSequence{T};
    error_rate
) -> BioSequences.LongSequence

Mycelia.open_fastx — Method

open_fastx(
    path::AbstractString
) -> Union{FASTX.FASTA.Reader, FASTX.FASTQ.Reader{T} where T<:(TranscodingStreams.TranscodingStream{C, S} where {S<:IO, C<:TranscodingStreams.Codec})}

Mycelia.open_genbank — Method

open_genbank(
    genbank_file
) -> Vector{GenomicAnnotations.Record}

Mycelia.open_gff — Method

open_gff(path::String) -> Any

Mycelia.parse_blast_report — Method

parse_blast_report(blast_report) -> DataFrames.DataFrame

Expects output type 7 from BLAST, default output type 6 doesn't have the header comments and won't auto-parse

Mycelia.parse_gfa — Method

parse_gfa(gfa) -> MetaGraphs.MetaGraph{Int64, Float64}

Mycelia.parse_jsonl — Method

parse_jsonl(filepath::String) -> Vector{Dict{String, Any}}

Mycelia.parse_mmseqs_easy_taxonomy_lca_tsv — Method

parse_mmseqs_easy_taxonomy_lca_tsv(
    lca_tsv
) -> DataFrames.DataFrame

Mycelia.parse_mmseqs_easy_taxonomy_tophit_report — Method

parse_mmseqs_easy_taxonomy_tophit_report(
    tophit_report
) -> DataFrames.DataFrame

Mycelia.parse_mmseqs_tophit_aln — Method

parse_mmseqs_tophit_aln(tophit_aln) -> DataFrames.DataFrame

Mycelia.parse_qualimap_contig_coverage — Method

parse_qualimap_contig_coverage(
    qualimap_report_txt
) -> DataFrames.DataFrame

Parse the contig coverage information from qualimap bamqc text report, which looks like the following:

# this is spades
>>>>>>> Coverage per contig

	NODE_1_length_107478_cov_9.051896	107478	21606903	201.0355886786133	60.39424208607496
	NODE_2_length_5444_cov_1.351945	5444	153263	28.152645113886848	5.954250612823136
	NODE_3_length_1062_cov_0.154390	1062	4294	4.043314500941619	1.6655384692688975
	NODE_4_length_776_cov_0.191489	776	3210	4.13659793814433	2.252009588980858

# below is megahit
>>>>>>> Coverage per contig

	k79_175	235	3862	16.43404255319149	8.437436249612457
	k79_89	303	3803	12.551155115511552	5.709975376279777
	k79_262	394	6671	16.931472081218274	7.579217802849293
	k79_90	379	1539	4.060686015831134	1.2929729111266581
	k79_91	211	3749	17.767772511848342	11.899185693011933
	k79_0	2042	90867	44.49902056807052	18.356525483516613

To make this more robust, consider reading in the names of the contigs from the assembled fasta

Mycelia.parse_rtg_eval_output — Method

parse_rtg_eval_output(f) -> DataFrames.DataFrame

Mycelia.parse_transterm_output — Method

parse_transterm_output(
    transterm_output
) -> DataFrames.DataFrame

Mycelia.parse_virsorter_score_tsv — Method

parse_virsorter_score_tsv(
    virsorter_score_tsv
) -> DataFrames.DataFrame

Mycelia.parse_xam_to_mapped_records_table — Function

parse_xam_to_mapped_records_table(
    xam
) -> DataFrames.DataFrame
parse_xam_to_mapped_records_table(
    xam,
    primary_only
) -> DataFrames.DataFrame

Mycelia.parse_xam_to_primary_mapping_table — Method

parse_xam_to_primary_mapping_table(
    xam
) -> DataFrames.DataFrame

Mycelia.parse_xam_to_summary_table — Method

parse_xam_to_summary_table(xam) -> DataFrames.DataFrame

Mycelia.parse_xam_to_taxonomic_mapping_quality — Function

parse_xam_to_taxonomic_mapping_quality(
    xam
) -> DataFrames.DataFrame
parse_xam_to_taxonomic_mapping_quality(
    xam,
    primary_only
) -> DataFrames.DataFrame

Mycelia.path_to_sequence — Method

path_to_sequence(kmers, path) -> Any

Mycelia.plot_graph — Method

plot_graph(graph) -> Any

Mycelia.plot_kmer_frequency_spectra — Method

plot_kmer_frequency_spectra(
    counts;
    log_scale,
    kwargs...
) -> Plots.Plot

Plots a histogram of kmer counts against # of kmers with those counts

Returns the plot object for adding additional layers and saving

Mycelia.plot_optimal_cluster_assessment_results — Method

plot_optimal_cluster_assessment_results(
    clustering_results
) -> Any

A short description of the function

julia> 1 + 1
2

Mycelia.polish_fastq — Method

polish_fastq(; fastq, k)

Mycelia.prefetch — Method

prefetch(; SRR, outdir)

Mycelia.process_fastq_record — Method

process_fastq_record(
;
    record,
    kmer_graph,
    yen_k_shortest_paths_and_weights,
    yen_k
)

Mycelia.q_value_to_error_rate — Method

q_value_to_error_rate(q_value) -> Any

Mycelia.random_fasta_record — Method

random_fasta_record(; seed, L) -> FASTX.FASTA.Record

Mycelia.rclone_copy — Method

rclone_copy(source, dest; config, max_attempts, sleep_timer)

Mycelia.rclone_list_directories — Method

rclone_list_directories(path) -> Any

Mycelia.read_fastani — Method

read_fastani(path::String) -> DataFrames.DataFrame

Imports results of fastani

Mycelia.read_gff — Method

read_gff(gff::AbstractString) -> DataFrames.DataFrame

Mycelia.read_gff — Method

read_gff(gff_io) -> DataFrames.DataFrame

Mycelia.read_kraken_report — Method

read_kraken_report(kraken_report) -> DataFrames.DataFrame

Mycelia.read_mmseqs_easy_search — Method

read_mmseqs_easy_search(mmseqs_file) -> DataFrames.DataFrame

Mycelia.reverse_translate — Method

reverse_translate(
    protein_sequence::BioSequences.LongAA
) -> BioSequences.LongSequence{BioSequences.DNAAlphabet{2}}

Mycelia.run_blast — Method

run_blast(
;
    out_dir,
    fasta,
    blast_db,
    blast_command,
    force,
    remote,
    wait
)

julia> 1 + 1
2

Mycelia.run_blastn — Method

run_blastn(
;
    out_dir,
    fasta,
    blast_db,
    task,
    force,
    remote,
    wait
)

julia> 1 + 1
2

Mycelia.run_clustal_omega — Method

Runs clustal omega on a fasta file

valid outfmts include

["fasta", "clustal", "msf", "phylip", "selex", "stockholm", "vienna"]

Mycelia.run_ectyper — Method

run_ectyper(fasta_file) -> Any

Mycelia.run_mlst — Method

run_mlst(fasta_file) -> String

Mycelia.run_mmseqs_easy_search — Method

run_mmseqs_easy_search(
;
    query_fasta,
    target_database,
    out_dir,
    outfile,
    format_output,
    threads,
    force
)

julia> 1 + 1
2

Mycelia.run_padloc — Method

run_padloc(fasta_file) -> Union{Nothing, Base.Process}

Run the 'padloc' tool from the 'padlocbio' conda environment on a given FASTA file.

https://doi.org/10.1093/nar/gkab883 https://github.com/padlocbio/padloc

This function first ensures that the 'padloc' environment is available via Bioconda. It then attempts to update the 'padloc' database. If a 'padloc' output file (with a '_padloc.csv' suffix) does not already exist for the input FASTA file, it runs 'padloc' with the specified FASTA file as input.

Mycelia.run_prodigal — Method

run_prodigal(; fasta_file, out_dir)

julia> 1 + 1
2

Mycelia.run_samtools_flagstat — Function

run_samtools_flagstat(xam) -> Any
run_samtools_flagstat(xam, samtools_flagstat) -> Any

Mycelia.run_transterm — Method

run_transterm(; fasta, gff)

Mycelia.samtools_index_fasta — Method

samtools_index_fasta(; fasta)

Mycelia.save_graph — Method

save_graph(
    graph::Graphs.AbstractGraph,
    outfile::String
) -> String

Mycelia.scg_sbatch — Method

scg_sbatch(
;
    job_name,
    mail_user,
    mail_type,
    logdir,
    partition,
    account,
    nodes,
    ntasks,
    time,
    cpus_per_task,
    mem_gb,
    cmd
)

Submit a command to SLURM using sbatch

Mycelia.seq2sha256 — Method

seq2sha256(seq::AbstractString) -> String

Mycelia.seq2sha256 — Method

seq2sha256(seq::BioSequences.BioSequence) -> String

Mycelia.sequence_to_stranded_path — Method

sequence_to_stranded_path(
    stranded_kmers,
    sequence
) -> Vector{Pair{Int64, Bool}}

Mycelia.setup_taxonkit_taxonomy — Method

setup_taxonkit_taxonomy() -> String

Mycelia.showall_blastdbs — Method

showall_blastdbs(; source) -> Any

Mycelia.simulate_nanopore_reads — Method

simulate_nanopore_reads(; fasta, quantity, outfile)

quantity should be either fold coverage (e.g. "50x"), or total bases sequenced (e.g. 1000000) - NOT TOTAL READS

See also: simulate_pacbio_reads, simulate_nearly_perfect_long_reads, simulate_short_reads

Mycelia.simulate_nearly_perfect_long_reads — Method

simulate_nearly_perfect_long_reads()

quantity should be either fold coverage (e.g. "50x"), or total bases sequenced (e.g. 1000000) - NOT TOTAL READS

See also: simulate_pacbio_reads, simulate_nanopore_reads, simulate_short_reads

Mycelia.simulate_pacbio_reads — Method

simulate_pacbio_reads(; fasta, quantity, outfile)

quantity should be either fold coverage (e.g. "50x"), or total bases sequenced (e.g. 1000000) - NOT TOTAL READS

Reads are ~ 15kb

See also: simulate_nanopore_reads, simulate_nearly_perfect_long_reads, simulate_short_reads

Mycelia.simulate_short_reads — Method

simulate_short_reads(; in_fasta, coverage, outbase)

See also: simulate_nanopore_reads, simulate_nearly_perfect_long_reads, simulate_pacbio_reads

Mycelia.simulate_variants — Method

simulate_variants(
    fasta_record::FASTX.FASTA.Record;
    n_variants,
    window_size,
    variant_size_disbribution,
    variant_type_likelihoods
) -> Any

Mycelia.simulate_variants — Method

simulate_variants(fasta_file::String) -> String

Mycelia.sort_fastq — Function

sort_fastq(input_fastq) -> String
sort_fastq(input_fastq, output_fastq) -> Any

This turns a 4-line FASTQ entry into a single tab separated line, adds a column with the length of each read, passes it to Unix sort, removes the length column, and converts it back into a FASTQ file.

sorts longest to shortest!!

http://thegenomefactory.blogspot.com/2012/11/sorting-fastq-files-by-sequence-length.html

Mycelia.subsample_reads_seqkit — Method

subsample_reads_seqkit(
;
    in_fastq,
    out_fastq,
    n_reads,
    proportion_reads
)

Mycelia.system_mem_to_minimap_index_size — Method

system_mem_to_minimap_index_size(
;
    system_mem_gb,
    denominator
)

Mycelia.system_overview — Method

system_overview(
;
    path
) -> @NamedTuple{system_threads::Int64, julia_threads::Int64, total_memory::String, available_memory::String, occupied_memory::String, total_storage::String, available_storage::String, occupied_storage::String}

Mycelia.tar_extract — Method

tar_extract(; tarchive, directory)

Mycelia.taxids2lca — Method

taxids2lca(ids::Vector{Int64}) -> Int64

Mycelia.taxids2ncbi_taxonomy_table — Method

taxids2ncbi_taxonomy_table(
    taxids::AbstractVector{Int64}
) -> DataFrames.DataFrame

Mycelia.taxonomic_id_to_children — Method

taxonomic_id_to_children(
    tax_id;
    DATABASE_ID,
    USERNAME,
    PASSWORD
)

Description

julia> 1 + 1
2

Mycelia.translate_nucleic_acid_fasta — Method

translate_nucleic_acid_fasta(
    fasta_nucleic_acid_file,
    fasta_amino_acid_file
) -> Any

Mycelia.transterm_output_to_gff — Method

transterm_output_to_gff(transterm_output) -> Any

Mycelia.trim_galore — Method

trim_galore(
;
    outdir,
    identifier
) -> Union{Nothing, Base.Process}

Mycelia.type_to_string — Method

type_to_string(T::AbstractString) -> Any

Mycelia.type_to_string — Method

type_to_string(T) -> Any

Mycelia.update_fasta_with_vcf — Method

update_fasta_with_vcf(; in_fasta, vcf_file, out_fasta)

Mycelia.update_gff_with_mmseqs — Method

update_gff_with_mmseqs(gff_file, mmseqs_file)

Mycelia.upload_edge_type_over_url_from_graph — Method

upload_edge_type_over_url_from_graph(
;
    src_type,
    dst_type,
    edge_type,
    graph,
    ADDRESS,
    USERNAME,
    PASSWORD,
    DATABASE,
    window_size
)

Mycelia.upload_node_over_api — Method

upload_node_over_api(
    graph,
    v;
    ADDRESS,
    USERNAME,
    PASSWORD,
    DATABASE
)

Description

julia> 1 + 1
2

Mycelia.upload_node_table — Method

upload_node_table(
;
    table,
    window_size,
    address,
    password,
    username,
    database,
    neo4j_import_dir
)

Mycelia.upload_node_type_over_url_from_graph — Method

upload_node_type_over_url_from_graph(
;
    node_type,
    graph,
    ADDRESS,
    USERNAME,
    PASSWORD,
    DATABASE,
    window_size
)

Mycelia.upload_nodes_over_api — Method

upload_nodes_over_api(
    graph;
    ADDRESS,
    USERNAME,
    PASSWORD,
    DATABASE
)

Mycelia.upload_nodes_to_neo4j — Method

upload_nodes_to_neo4j(
;
    graph,
    address,
    username,
    password,
    format,
    database,
    neo4j_import_directory
)

Mycelia.vcat_with_missing — Method

vcat_with_missing(
    dfs::DataFrames.AbstractDataFrame...
) -> Union{DataFrames.DataFrame, Vector{Any}}

Mycelia.viterbi_maximum_likelihood_traversals — Method

viterbi_maximum_likelihood_traversals(
    stranded_kmer_graph;
    error_rate,
    verbosity
) -> Vector{FASTX.FASTA.Record}

Mycelia.wcss — Method

wcss(clustering_result) -> Any

A short description of the function

julia> 1 + 1
2

Mycelia.write_fasta — Method

write_fasta(; outfile, records, gzip)

Writes FASTA records to a file, optionally gzipped.

Arguments

outfile::AbstractString: Path to the output FASTA file. Will append ".gz" if gzip is true.

records::Vector{FASTX.FASTA.Record}: A vector of FASTA records.

gzip::Bool=false: Whether to compress the output with gzip.

Returns

outfile::String: The path to the output FASTA file (including ".gz" if applicable).

Mycelia.write_gff — Method

write_gff(; gff, outfile)

Mycelia.write_vcf_table — Method

write_vcf_table(; vcf_file, vcf_table, fasta_file)

Mycelia.xam_to_contig_mapping_stats — Method

xam_to_contig_mapping_stats(xam) -> Any