Docstrings

Mycelia.assess_alignment

Mycelia.assess_alignment_accuracy

Mycelia.assess_optimal_alignment

Mycelia.count_canonical_kmers

Mycelia.count_kmers

Mycelia.determine_edge_probabilities

Mycelia.determine_edge_probabilities

Mycelia.find_downstream_vertices

Mycelia.find_unbranched_neighbors

Mycelia.get_gff

Mycelia.get_kmer_index

Mycelia.get_sequence

Mycelia.is_equivalent

Mycelia.observe

Mycelia.observe

Mycelia.open_fastx

Mycelia.open_gff

Mycelia.oriented_unbranching_walk

Mycelia.plot_kmer_frequency_spectra

Mycelia.set_metadata!

Mycelia.set_metadata!

Mycelia.assess_alignment — Method

assess_alignment(a, b) -> NamedTuple{(:total_matches, :total_edits), Tuple{Int64, Int64}}

A short description of the function

julia> 1 + 1
2

Mycelia.assess_alignment_accuracy — Method

assess_alignment_accuracy(alignment_result) -> Any

Return proportion of matched bases in alignment to total matches + edits.

0-1, not %

julia> 1 + 1
2

Mycelia.assess_optimal_alignment — Method

assess_optimal_alignment(kmer, observed_kmer) -> Tuple{NamedTuple{(:total_matches, :total_edits), Tuple{Int64, Int64}}, Union{Missing, Bool}}

Used to determine which orientation provides an optimal alignment for initiating path likelihood analyses in viterbi analysis

julia> 1 + 1
2

Mycelia.count_canonical_kmers — Method

count_canonical_kmers(_::Type{KMER_TYPE}, sequence::BioSequences.LongSequence) -> OrderedCollections.OrderedDict{_A, Int64} where _A

A short description of the function

julia> 1 + 1
2

Mycelia.count_kmers — Method

A short description of the function

julia> 1 + 1
2

Mycelia.determine_edge_probabilities — Method

determine_edge_probabilities(graph, strand) -> Union{SparseArrays.SparseMatrixCSC{Float64, Int64}, SparseArrays.SparseVector{_A, Int64} where _A}

A short description of the function

julia> 1 + 1
2

Mycelia.determine_edge_probabilities — Method

determine_edge_probabilities(graph) -> Tuple{Union{SparseArrays.SparseMatrixCSC{Float64, Int64}, SparseArrays.SparseVector{_A, Int64} where _A}, Union{SparseArrays.SparseMatrixCSC{Float64, Int64}, SparseArrays.SparseVector{_A, Int64} where _A}}

A short description of the function

julia> 1 + 1
2

Mycelia.find_downstream_vertices — Method

find_downstream_vertices(kmer_graph, vertex, orientation) -> Set{Int64}

Get dna (db = "nuccore") or protein (db = "protein") sequences from NCBI or get fasta directly from FTP site

julia> 1 + 1
2

Mycelia.find_unbranched_neighbors — Method

find_unbranched_neighbors(kmer_graph, vertex, orientation) -> Union{Set{Int64}, Vector{Int64}}

Get dna (db = "nuccore") or protein (db = "protein") sequences from NCBI or get fasta directly from FTP site

julia> 1 + 1
2

Mycelia.get_gff — Method

get_gff(; db, accession, ftp) -> Union{Nothing, GFF3.Reader}

Get dna (db = "nuccore") or protein (db = "protein") sequences from NCBI or get fasta directly from FTP site

julia> 1 + 1
2

Mycelia.get_kmer_index — Method

get_kmer_index(kmers, kmer) -> Any

A short description of the function

julia> 1 + 1
2

Mycelia.get_sequence — Method

get_sequence(; db, accession, ftp) -> Union{Nothing, FASTX.FASTA.Reader}

Get dna (db = "nuccore") or protein (db = "protein") sequences from NCBI or get fasta directly from FTP site

julia> 1 + 1
2

Mycelia.is_equivalent — Method

is_equivalent(a, b) -> Any

A short description of the function

julia> 1 + 1
2

Mycelia.observe — Method

observe(record::Union{FASTX.FASTA.Record, FASTX.FASTQ.Record}; error_rate) -> FASTX.FASTQ.Record

A short description of the function

julia> 1 + 1
2

Mycelia.observe — Method

observe(sequence::BioSequences.LongSequence{T}; error_rate) -> BioSequences.LongSequence

A short description of the function

julia> 1 + 1
2

Mycelia.open_fastx — Method

open_fastx(path::String) -> Any

A short description of the function

julia> 1 + 1
2

Mycelia.open_gff — Method

open_gff(path::String) -> GFF3.Reader

A short description of the function

julia> 1 + 1
2

Mycelia.oriented_unbranching_walk — Method

oriented_unbranching_walk(kmer_graph, vertex, orientation) -> Vector{Any}

Get dna (db = "nuccore") or protein (db = "protein") sequences from NCBI or get fasta directly from FTP site

julia> 1 + 1
2

Mycelia.plot_kmer_frequency_spectra — Method

plot_kmer_frequency_spectra(counts; log_scale, kwargs...) -> Plots.Plot

A short description of the function

julia> 1 + 1
2

Mycelia.set_metadata! — Method

set_metadata!(kmer_graph, edge::Graphs.SimpleGraphs.SimpleEdge, key, value) -> Bool

A short description of the function

julia> 1 + 1
2

Mycelia.set_metadata! — Method

set_metadata!(kmer_graph, vertex::Integer, key, value) -> Bool

A short description of the function

julia> 1 + 1
2